Added HMI, Adjusted PID logic
deleted base motor speed so that offset directly changes the speed of the motors instead of calculating the offset between the two motors. added encoder to modify the pid params, use button to switch which param to modify. a singular function modifies the param values and display them. added some comments